# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Maeglagi API
  version: 0.1.0
paths:
  /api/v1/health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: API health
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /api/v1/workspaces/{workspaceId}/graph:
    get:
      operationId: getKnowledgeGraph
      summary: Knowledge graph of a workspace, as of a moment in time
      description: >
        Nodes are entities, edges are relations valid at `at` (default now). A relation with no
        stated start or end counts as open. A decision replaced by a later one has `supersededBy`
        and is joined to its successor by a `supersedes` edge (successor -> replaced).
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
        - name: at
          in: query
          required: false
          description: ISO 8601 instant. Defaults to now.
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: The graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeGraph"
        "404": { description: Workspace not found }
        "503": { description: Graph store is not configured }
  /api/v1/workspaces/{workspaceId}/ask:
    post:
      operationId: ask
      summary: Answer a question from the workspace's sources, streaming, with its evidence
      description: >
        Server-sent events, one `data: {json}` line per event, ending with `data: [DONE]`.
        Order is `sources`, then `token` events, then `done`. When nothing relevant is found the
        stream is a single `error` event and no model is called. A failure after the stream has
        started is also reported as an `error` event. Citations in the text (`[1]`) always refer
        to an `index` in the `sources` event; citations to missing evidence are removed.
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        "200":
          description: Event stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/AnswerEvent"
        "404": { description: Workspace not found }
        "422": { description: Blank question, or the workspace has no key that can chat or embed }
  /api/v1/llm-keys/models:
    post:
      operationId: listKeyModels
      summary: Models a key can use, sorted into jobs, before a workspace exists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider]
              properties:
                provider: { type: string }
                credentialId: { type: string, format: uuid, description: Preview a saved account credential }
                apiKey: { type: string, writeOnly: true, description: Optional for Ollama }
                baseUrl: { type: string, description: Required for Ollama }
      responses:
        "200":
          description: One entry per job, always in the order answer, extraction, embedding, transcription
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceModels" }
        "422": { description: The key is not valid }
  /api/v1/provider-credentials:
    get:
      operationId: listAccountProviderCredentials
      summary: List the signed-in account's saved provider connections
      responses:
        "200":
          description: Credential metadata without secrets
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ProviderCredential" } }
    post:
      operationId: addAccountProviderCredential
      summary: Save a provider connection for reuse in workspaces
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialInput" }
      responses:
        "201":
          description: Saved credential metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderCredential" }
  /api/v1/provider-credentials/{credentialId}:
    parameters:
      - { name: credentialId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      operationId: updateAccountProviderCredential
      summary: Update a saved account credential by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                apiKey: { type: string, writeOnly: true }
                baseUrl: { type: string }
      responses:
        "200":
          description: Updated metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderCredential" }
    delete:
      operationId: deleteAccountProviderCredential
      summary: Remove an account credential unused by workspace model choices
      responses:
        "204": { description: Removed }
        "409": { description: Credential is selected by a workspace }
  /api/v1/provider-credentials/{credentialId}/default:
    parameters:
      - { name: credentialId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      operationId: chooseAccountDefaultProviderCredential
      summary: Mark an account credential as the default
      responses:
        "200":
          description: Updated metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderCredential" }
  /api/v1/workspaces/{workspaceId}/provider-credentials:
    parameters:
      - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: listProviderCredentials
      summary: List credential metadata without secrets
      responses:
        "200":
          description: Credentials ordered by creation time
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ProviderCredential" } }
        "404": { description: Workspace not found }
    post:
      operationId: addProviderCredential
      summary: Add a labeled key without changing the workspace default or model choices
      description: Provider and label must be unique within the workspace. The new key is not the default.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialInput" }
      responses:
        "201":
          description: Credential metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderCredential" }
        "404": { description: Workspace not found }
        "409": { description: Provider and label already exist }
        "422": { description: Provider or key is invalid }
  /api/v1/workspaces/{workspaceId}/provider-credentials/{credentialId}:
    parameters:
      - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: credentialId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      operationId: rotateProviderCredential
      summary: Update one credential's key or Ollama address without changing model choices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                apiKey: { type: string, writeOnly: true, description: Omission preserves Ollama secret; empty clears it }
                baseUrl: { type: string, description: Omission preserves Ollama address }
      responses:
        "200":
          description: Updated credential metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderCredential" }
        "404": { description: Workspace or credential not found }
        "422": { description: Key is invalid }
    delete:
      operationId: removeProviderCredential
      summary: Remove one credential and its Vault secret
      description: >
        If other credentials exist, first choose another default before deleting the default.
        The last active credential for a provider selected by a workspace model cannot be
        removed. Removing the final credential is allowed when no model selects its provider;
        the workspace then has no default. Model choices never change during deletion.
      responses:
        "204": { description: Credential and Vault secret removed }
        "404": { description: Workspace or credential not found }
        "409": { description: Default must be changed first or a selected model needs this provider }
  /api/v1/workspaces/{workspaceId}/provider-credentials/{credentialId}/default:
    parameters:
      - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: credentialId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      operationId: chooseDefaultProviderCredential
      summary: Select the workspace default key without changing model choices
      responses:
        "200":
          description: Selected credential metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderCredential" }
        "404": { description: Workspace or credential not found }
        "409": { description: Credential is inactive }
  /api/v1/workspaces/{workspaceId}/llm-key:
    parameters:
      - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: getDefaultLlmKey
      summary: Get the default credential metadata, or null
      responses:
        "200":
          description: Default credential metadata or null
          content:
            application/json:
              schema:
                oneOf: [{ $ref: "#/components/schemas/ProviderCredential" }, { type: "null" }]
        "404": { description: Workspace not found }
    put:
      operationId: upsertDefaultLlmKey
      summary: Legacy key save, rotating the matching provider and label and making it default
      description: Preserves the existing frontend flow and leaves model choices unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialInput" }
      responses:
        "200":
          description: Default credential metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderCredential" }
        "404": { description: Workspace not found }
        "422": { description: Provider or key is invalid }
  /api/v1/workspaces/{workspaceId}/ai/models:
    parameters:
      - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: getWorkspaceModels
      summary: Models the workspace's keys offer per job, and which one is chosen
      responses:
        "200":
          description: The models
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceModels" }
        "404": { description: Workspace not found }
    put:
      operationId: updateWorkspaceModels
      summary: Choose the model used for each job
      description: Jobs left out keep their current choice. LLM jobs (answer, extraction, transcription) change freely. The embedding model is fixed when the workspace is created, so changing it is a 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [selections]
              properties:
                selections:
                  type: object
                  additionalProperties: { $ref: "#/components/schemas/ModelOption" }
      responses:
        "200":
          description: The models after the change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceModels" }
        "404": { description: Workspace not found }
        "409": { description: The embedding model is fixed at workspace creation and cannot be changed }
        "422": { description: A job is unknown or the model is not offered for it }
components:
  schemas:
    CredentialInput:
      type: object
      required: [provider]
      properties:
        provider: { type: string }
        apiKey: { type: string, writeOnly: true, description: Optional for Ollama }
        baseUrl: { type: string, description: Required for new Ollama connections }
        label: { type: string, minLength: 1, maxLength: 80, default: "기본" }
    ProviderCredential:
      type: object
      required: [id, provider, label, baseUrl, keyHint, status, isDefault, updatedAt]
      properties:
        id: { type: string, format: uuid }
        provider: { type: string }
        label: { type: string }
        baseUrl: { type: [string, "null"] }
        keyHint: { type: string }
        status: { type: string }
        isDefault: { type: boolean }
        updatedAt: { type: string, format: date-time }
    ModelOption:
      type: object
      required: [provider, model]
      properties:
        provider: { type: string }
        model: { type: string }
        credentialId: { type: string, format: uuid, description: Identifies the saved account connection for workspace options }
    RoleModels:
      type: object
      required: [role, options, selected, locked]
      properties:
        role: { type: string, enum: [answer, extraction, embedding, transcription] }
        options:
          type: array
          description: Ordered lighter models first, stable aliases before dated snapshots. Not a quality ranking.
          items: { $ref: "#/components/schemas/ModelOption" }
        selected:
          oneOf: [{ $ref: "#/components/schemas/ModelOption" }, { type: "null" }]
        locked: { type: boolean }
    WorkspaceModels:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          items: { $ref: "#/components/schemas/RoleModels" }
    AnswerEvent:
      oneOf:
        - type: object
          required: [type, sources]
          properties:
            type: { type: string, const: sources }
            sources:
              type: array
              items: { $ref: "#/components/schemas/AnswerSource" }
        - type: object
          required: [type, text]
          properties:
            type: { type: string, const: token }
            text: { type: string }
        - type: object
          required: [type]
          properties:
            type: { type: string, const: done }
        - type: object
          required: [type, message]
          properties:
            type: { type: string, const: error }
            message: { type: string }
    AnswerSource:
      type: object
      required: [index, sourceId, chunkId, kind, title, excerpt]
      properties:
        index:
          type: integer
          description: The n in the answer's [n] citations
        sourceId: { type: string, format: uuid }
        chunkId: { type: string, format: uuid }
        kind: { type: string, enum: [document, meeting] }
        title: { type: string }
        excerpt: { type: string }
        timestamp:
          type: number
          description: Seconds into the meeting where this evidence starts. Absent for documents.
    KnowledgeGraph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items: { $ref: "#/components/schemas/GraphNode" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/GraphEdge" }
    GraphNode:
      type: object
      required: [id, type, kind, label, degree, sources]
      properties:
        id: { type: string }
        type:
          type: string
          enum: [person, project, decision, task, event]
          description: Frontend entity type. Organization maps to person; Technology to project; Meeting, Event, Issue and Document to event.
        kind:
          type: string
          enum: [Person, Project, Organization, Meeting, Document, Issue, Task, Decision, Event, Technology]
        label: { type: string }
        degree: { type: integer, description: Number of edges drawn at `at` }
        sources:
          type: array
          items: { $ref: "#/components/schemas/GraphSource" }
        supersededBy:
          type: string
          description: Id of the decision that explicitly replaced this decision
    GraphSource:
      type: object
      required: [id, kind, title]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        title: { type: string }
    GraphEdge:
      type: object
      required: [id, source, target, type]
      properties:
        id: { type: string }
        source: { type: string }
        target: { type: string }
        type:
          type: string
          enum: [participates_in, decided, assigned_to, blocks, relates_to, supersedes]
          description: BLOCKED_BY is drawn reversed (blocker -> blocked). Relations with no frontend equivalent are relates_to.
        kind:
          type: string
          description: Exact ontology relation, or SUPERSEDES for a supersedes edge
        validFrom: { type: string, format: date-time }
        validTo: { type: string, format: date-time }
    Health:
      type: object
      required: [status, service, version]
      properties:
        status: { type: string, const: ok }
        service: { type: string }
        version: { type: string }